requirements.txt
pycryptodome>=3.20
flask>=3.0
requests>=2.31
